requirements.txt
asgiref==3.7.2
click==8.1.6
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
joblib==1.3.1
nltk==3.8.1
numpy==1.25.2
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
regex==2023.6.3
scikit-learn==1.2.2
scipy==1.11.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
